PL_destructcb: A C callback invoked after all Perl code is done
PL_destructcb is of type "void (*fun)(pTHX)". If not NULL, it is invoked
close to the very end of the interpreter destruction, after all Perl
code is done and not even late DESTROY methods may call back into Perl
land. This is for XS modules that need to clean up after themselves
avoiding any chance of segfaulting when calling back into Perl code.