(perl #127663) use link when rename() isn't available
I suspect this is unnecessary, since unrelated tests failed without
rename(), but it's not a huge change.

Also, if rename() is available, attempt use link() to create the backup
so there's always some version of the original file.  Falls back to
rename() if the link() fails, since the filesystem might not support
link().